+++ packages/react/src/dynamic/index.ts
@@ -1,5 +1,3 @@
-'use client'
-
 export { TokenIcon } from './TokenIcon'
 export { NetworkIcon } from './NetworkIcon'
 export { WalletIcon } from './WalletIcon'